backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
Pillow==10.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
